remove a file
