--- src/move/handle_keys.c
@@ -1,49 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   handle_keys.c                                      :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: caide-so <caide-so@student.42sp.org.br>    +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/07/21 20:53:13 by caide-so          #+#    #+#             */
-/*   Updated: 2025/07/21 20:53:29 by caide-so         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "../../include/cub3d.h"
-
-int	handle_key_press(int keycode, t_game *game)
-{
-	if (keycode == KEY_ESC)
-		exit_game(game);
-	else if (keycode == KEY_W)
-		game->keys.w = 1;
-	else if (keycode == KEY_S)
-		game->keys.s = 1;
-	else if (keycode == KEY_A)
-		game->keys.a = 1;
-	else if (keycode == KEY_D)
-		game->keys.d = 1;
-	else if (keycode == KEY_LEFT)
-		game->keys.left = 1;
-	else if (keycode == KEY_RIGHT)
-		game->keys.right = 1;
-	return (0);
-}
-
-int	handle_key_release(int keycode, t_game *game)
-{
-	if (keycode == KEY_W)
-		game->keys.w = 0;
-	else if (keycode == KEY_S)
-		game->keys.s = 0;
-	else if (keycode == KEY_A)
-		game->keys.a = 0;
-	else if (keycode == KEY_D)
-		game->keys.d = 0;
-	else if (keycode == KEY_LEFT)
-		game->keys.left = 0;
-	else if (keycode == KEY_RIGHT)
-		game->keys.right = 0;
-	return (0);
-}
